backend/requirements.txt
# Flask Backend Requirements
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Limiter==3.10.0
python-dotenv==1.0.1
requests==2.32.3
waitress==3.0.2

# Security, serving, and monitoring
Werkzeug==3.1.3
cryptography==42.0.5
CherryPy==18.9.0

# File parsing for agent knowledge
pdfplumber==0.11.0
python-docx==1.1.0

# Universal Document Tools - Core
PyPDF2==3.0.1
pypdf==5.1.0
mammoth==1.8.0
openpyxl==3.1.5
pandas==2.2.3
xlsxwriter==3.2.0
Pillow==10.4.0
reportlab==4.2.5

# CRITICAL: NumPy version constraint for PyTorch compatibility
# NumPy 2.x causes "Failed to initialize NumPy: _ARRAY_API not found" error with PyTorch 2.2.0
# Must use NumPy < 2.0 for compatibility
numpy<2.0,>=1.24.0

# Local Offline OCR - Requires NumPy < 2.0
# EasyOCR - Supports 80+ languages, runs completely offline
easyocr==1.7.1
torch==2.2.0
torchvision==0.17.0

# Optional: For advanced OCR features
# transformers - For EffOCR-Word and other transformer-based OCR
# pdf2image - For converting PDF pages to images for OCR
# transformers>=4.36.0
# pdf2image>=1.16.3

# Database ORM
SQLAlchemy==2.0.36

# LangGraph for Plan-Execute-Verify orchestration state machine
langgraph==0.2.62
langchain-core==0.3.27
pydantic==2.10.5

# Optional: Redis for distributed rate limiting
# redis==5.2.1
